Created notify() and introduced contacts.example
The notify function now allows the bot to notify users with a random quote. I will continue building this functionality as time goes on.

The 'contacts.example' script should be renamed to contacts.py.

TODO: Add a help tip for -n/--notify so that the user understands what specificity is required.